% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{calcBaHa}
\alias{calcBaHa}
\title{Calculate Basal Area}
\usage{
calcBaHa(trees, plotArea)
}
\arguments{
\item{trees}{Data frame with columns plotID and ba (basal area of each tree in m2).}

\item{plotArea}{Numeric value representing the plot area in m2.}
}
\value{
Data frame with plotID and total basal area (in m2 ha-1).
}
\description{
Computes the total basal area of trees in each plot.
}
\examples{
calcBa(trees, plotArea = 400)
}
