% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{calcHdom}
\alias{calcHdom}
\title{Calculate Dominant Height}
\usage{
calcHdom(trees, plotArea)
}
\arguments{
\item{trees}{Data frame with columns plotID, dbh (diameter at breast height), and h (height).}

\item{plotArea}{Numeric value representing the plot area in m??.}
}
\value{
Data frame with plotID and the mean dominant height (Hdom).
}
\description{
Computes the mean height of the dominant trees in each plot.
}
\examples{
calcHdom(trees, plotArea = 1000)
}
