% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{calcNHa}
\alias{calcNHa}
\title{Calculate Tree Density}
\usage{
calcNHa(trees, plotArea)
}
\arguments{
\item{trees}{Data frame with columns plotID and n (number of trees).}

\item{plotArea}{Numeric value representing the plot area in m??.}
}
\value{
Data frame with plotID and tree density (N).
}
\description{
Computes the density of trees in each plot, expressed as the number of trees per hectare.
}
\examples{
calcN(trees, plotArea = 250)
}
