% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{calcPlotSummaries}
\alias{calcPlotSummaries}
\title{Calculate plot summaries}
\usage{
calcPlotSummaries(trees)
}
\arguments{
\item{trees}{Data frame containing plotID, height of each tree ("h" in m) basal area ("ba" in m2), and volume ("vol" in m3).}
}
\value{
A data frame with plot-level summaries including the following columns:
- `plotID`: Identifier for each plot.
- `N`: Tree density per hectare.
- `Hlor`: Lorey's height.
- `Hdom`: Dominant height.
- `Vol`: Volume per hectare.
- `Ba`: Basal area per hectare.
}
\description{
This function computes summaries for each plot of number of trees per ha, Lorey's height (Hlor), dominant height, volume per hectare, and basal area per hectare.
}
\examples{
result <- calcPlotSummaries(trees)

}
\seealso{
calcNHa, calcHlor, calcHdom, calcVolHa, calcBaHa
}
