% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{calcVolHa}
\alias{calcVolHa}
\title{Calculate Volume per ha}
\usage{
calcVolHa(trees, plotArea)
}
\arguments{
\item{trees}{Data frame with columns plotID and vol (volume of each tree in m3).}

\item{plotArea}{Numeric value representing the plot area in m2.}
}
\value{
Data frame with total volume (m3 ha-1) for each plotID.
}
\description{
Computes the total volume of trees in each plot.
}
\examples{
calcVolHa(trees, plotArea = 400)
}
