% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{dbh2ba}
\alias{dbh2ba}
\title{Calculate Basal Area (BA)}
\usage{
dbh2ba(dbh)
}
\arguments{
\item{dbh}{Numeric vector of diameter at breast height in cm.}
}
\value{
Numeric vector of basal areas in square centimeters.
}
\description{
Computes the basal area of trees given diameters at breast height (dbh).
}
\examples{
calcBa(30)
}
