% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{fitHD_Plot}
\alias{fitHD_Plot}
\title{Fit Height-Diameter Model by Plot}
\usage{
fitHD_Plot(trees)
}
\arguments{
\item{trees}{Data frame with columns plotID, h (height in m), and dbh (diameter at breast height in cm).}
}
\value{
A fitted mixed-effects nonlinear model.
#' @references
Sharma, R. P.and Breidenbach, J. (2015). Modeling height-diameter relationships for Norway spruce, Scots pine, and downy birch using Norwegian national forest inventory data. Forest Science and Technology, 11(1), 44-53.
}
\description{
Fits a height-diameter model to each plot, allowing for random effects.
}
\examples{
fitHD_Plot(trees)
}
